+++ packages/integration/src/index.ts
@@ -6,6 +6,7 @@ export {
 export { getSourceFromVirtualCssFile } from './virtualFile';
 export { getPackageInfo } from './packageInfo';
 export { compile, vanillaExtractTransformPlugin } from './compile';
+export { createCompiler } from './compiler';
 export { hash } from './hash';
 export { addFileScope } from './addFileScope';
 export { serializeCss, deserializeCss } from './serialize';
@@ -16,3 +17,4 @@ export * from './filters';
 export type { IdentifierOption } from './types';
 export type { PackageInfo } from './packageInfo';
 export type { CompileOptions } from './compile';
+export type { Compiler, CreateCompilerOptions } from './compiler';