+++ test-helpers/src/startFixture/esbuild.ts
@@ -2,12 +2,13 @@ import path from 'path';
 import { existsSync, promises as fs } from 'fs';
 
 import { vanillaExtractPlugin } from '@vanilla-extract/esbuild-plugin';
-import { serve } from 'esbuild';
+import { vanillaExtractPlugin as vanillaExtractPluginNext } from '@vanilla-extract/esbuild-plugin-next';
+import * as esbuild from 'esbuild';
 
 import { TestServer } from './types';
 
 export interface EsbuildFixtureOptions {
-  type: 'esbuild' | 'esbuild-runtime';
+  type: 'esbuild' | 'esbuild-runtime' | 'esbuild-next' | 'esbuild-next-runtime';
   mode?: 'development' | 'production';
   port: number;
 }
@@ -15,6 +16,9 @@ export const startEsbuildFixture = async (
   fixtureName: string,
   { type, mode = 'development', port = 3000 }: EsbuildFixtureOptions,
 ): Promise<TestServer> => {
+  const plugin = type.includes('next')
+    ? vanillaExtractPluginNext
+    : vanillaExtractPlugin;
   const entry = require.resolve(`@fixtures/${fixtureName}/src/index.ts`);
   const absWorkingDir = path.dirname(
     require.resolve(`@fixtures/${fixtureName}/package.json`),
@@ -27,23 +31,22 @@ export const startEsbuildFixture = async (
 
   await fs.mkdir(outdir, { recursive: true });
 
-  const server = await serve(
-    { servedir: outdir, port },
-    {
-      entryPoints: [entry],
-      metafile: true,
-      platform: 'browser',
-      bundle: true,
-      minify: mode === 'production',
-      plugins: [
-        vanillaExtractPlugin({
-          runtime: type === 'esbuild-runtime',
-        }),
-      ],
-      absWorkingDir,
-      outdir,
-    },
-  );
+  const ctx = await esbuild.context({
+    entryPoints: [entry],
+    metafile: true,
+    platform: 'browser',
+    bundle: true,
+    minify: mode === 'production',
+    plugins: [
+      plugin({
+        runtime: type.includes('runtime'),
+      }),
+    ],
+    absWorkingDir,
+    outdir,
+  });
+
+  const server = await ctx.serve({ servedir: outdir, port });
 
   await fs.writeFile(
     path.join(outdir, 'index.html'),
@@ -66,7 +69,7 @@ export const startEsbuildFixture = async (
     url: `http://${server.host}:${port}`,
     stylesheet: 'index.css',
     close: () => {
-      server.stop();
+      ctx.dispose();
 
       return Promise.resolve();
     },